Add a generic add tool page
