+++ app/web_ui/src/routes/(app)/settings/manage_tools/[project_id]/add_tools/remote_mcp/+page.ts
@@ -1,0 +1,1 @@
+export const prerender = false
